+++ CHANGELOG.md
@@ -4,8 +4,9 @@
 
 - phishing 0.15.1
 - hw-ledger, hw-ledger-transports, keyring, networks, util, util-crypto, x-bigint, x-fetch, x-global, x-randomvalues, x-textdecoder, x-textencoder, x-ws 9.6.1
-- hw-ledger, hw-ledger-transports, keyring, networks, util, util-crypto, x-bigint, x-fetch, x-global, x-randomvalues, x-textdecoder, x-textencoder, x-ws 9.5.2-19
 - wasm-bridge, wasm-crypto, wasm-crypto-asmjs, wasm-crypto-init, wasm-crypto-wasm, wasm-util 6.1.5
+- Auto-create `mod.ts` on release from available root folders
+
 
 ## 0.0.0-8 Jun 23, 2022